asasa

+++ quizviran/Http/Controllers/RoomController.php
@@ -7,25 +7,25 @@ use App\Models\Comment;
 use Quizviran\Models\Room;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
-use Quizviran\Http\Resources\QuizResourse;
 use Morilog\Jalali\Jalalian;
 
 class RoomController extends Controller
 {
-    public function __construct()
+    public function create()
     {
         if (! auth()->check()) {
             return abort(404);
         }
-    }
 
-    public function create()
-    {
         return view('Quizviran::panel.teacher.roomCreate');
     }
 
     public function show($room)
     {
+        if (! auth()->check()) {
+            return abort(404);
+        }
+
         $room = Room::where('link', $room)->with([
             'comments',
             'comments.files',
@@ -44,6 +44,10 @@ class RoomController extends Controller
 
     public function store()
     {
+        if (! auth()->check()) {
+            return abort(404);
+        }
+
         if (auth()->user()->type != 'teacher') {
             return back();
         }
@@ -70,6 +74,10 @@ class RoomController extends Controller
 
     public function addComment(Request $request)
     {
+        if (! auth()->check()) {
+            return abort(404);
+        }
+
         $files = collect($request->post('files'));
         $files = File::find($files->pluck('id'));
         $comment = Comment::create([
@@ -86,6 +94,10 @@ class RoomController extends Controller
 
     public function updateComment(Comment $comment, Request $request)
     {
+        if (! auth()->check()) {
+            return abort(404);
+        }
+
         if (! $comment->isOwn()) {
             return response(['message' => 'این نظر قابل ویرایش نیست.'], 400);
         }
@@ -97,6 +109,10 @@ class RoomController extends Controller
 
     public function deleteComment(Comment $comment)
     {
+        if (! auth()->check()) {
+            return abort(404);
+        }
+
         if ($comment->isOwn() || $comment->isOwnMember()) {
             \Illuminate\Support\Facades\File::delete(public_path($comment->files->pluck('file')));
             $comment->delete();
@@ -109,6 +125,10 @@ class RoomController extends Controller
 
     public function members($room)
     {
+        if (! auth()->check()) {
+            return abort(404);
+        }
+
         $room = Room::where('link', $room)->with('members')->firstOrFail();
         $room->created_at = Jalalian::forge($room->created_at);
 
@@ -121,6 +141,10 @@ class RoomController extends Controller
 
     public function lock($room)
     {
+        if (! auth()->check()) {
+            return abort(404);
+        }
+
         $room = Room::findOrFail($room);
         if(!(auth()->user()->hasRoom($room) && auth()->user()->type == 'teacher'))
             return back();
